:sparkles: feat : 투표 / 취소 시 각 횟수만큼 펫 조회모델 업데이트

+++ src/main/java/numble/pet/vote/pet/query/domain/PetData.java
@@ -45,11 +45,11 @@ public class PetData implements Serializable {
   @Field(name = "updated_at")
   private LocalDateTime updatedAt;
 
-  public void addVoteCount(Long count) {
+  public void addVoteCount(int count) {
     this.voteCount += count;
   }
 
-  public void subtractVoteCount(Long count) {
+  public void subtractVoteCount(int count) {
     this.voteCount -= count;
   }
